feat: 최근 입장한방 구현

+++ app/src/main/java/umc/onairmate/data/api/UserService.kt
@@ -1,10 +1,10 @@
 package umc.onairmate.data.api
 
-import retrofit2.http.GET
 import retrofit2.http.DELETE
+import retrofit2.http.GET
 import retrofit2.http.Header
 import retrofit2.http.Path
-import umc.onairmate.data.model.response.ParticipatedRoomData
+import umc.onairmate.data.model.entity.ParticipatedRoomData
 import umc.onairmate.data.model.response.MessageResponse
 import umc.onairmate.data.model.response.RawDefaultResponse
 
@@ -17,7 +17,7 @@ interface UserService {
     @DELETE("users/participated-rooms/{roomId}")
     suspend fun deleteParticipatedRoom(
         @Header("Authorization") accessToken: String, // "Bearer xxx"
-        @Path("roomId") roomId: Long
+        @Path("roomId") roomId: Int
     ): RawDefaultResponse<MessageResponse>
 
 }
